# src/app/api/rooms/join/[code]/route.ts
import { NextResponse } from 'next/server';
import prisma from '@/lib/db';

export async function GET(
  _request: Request,
  { params }: { params: Promise<{ code: string }> }
) {
  try {
    const { code } = await params;

    const room = await prisma.room.findUnique({
      where: { code: code.toUpperCase() },
      select: {
        id: true,
        name: true,
        code: true,
        isActive: true,
        students: {
          where: { isRemoved: false },
          select: {
            id: true,
            name: true,
            seatNumber: true,
          },
          orderBy: [{ seatNumber: 'asc' }, { name: 'asc' }],
        },
        items: {
          where: { isActive: true },
          select: {
            id: true,
            name: true,
            dueDate: true,
          },
          orderBy: { createdAt: 'desc' },
        },
      },
    });

    if (!room) {
      return NextResponse.json(
        { error: '找不到該房間，請確認代碼是否正確' },
        { status: 404 }
      );
    }

    if (!room.isActive) {
      return NextResponse.json(
        { error: '該房間已停用' },
        { status: 404 }
      );
    }

    return NextResponse.json({
      room: {
        id: room.id,
        name: room.name,
        code: room.code,
      },
      students: room.students,
      items: room.items,
    });
  } catch (error) {
    console.error('Failed to join room:', error);
    return NextResponse.json(
      { error: '加入房間失敗' },
      { status: 500 }
    );
  }
}